chore: refactor example code
