+++ example/lib/core/constants.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+const loremText = 'Lorem ipsum dolor sit amet,\nconsectetur adipiscing elit.';
+
+const demoTextStyle = TextStyle(
+  fontSize: 48,
+  fontWeight: FontWeight.w800,
+  color: Color(0xFF1E293B),
+  height: 1.2,
+  letterSpacing: -0.5,
+);
+
+const letterAnimationDuration = Duration(milliseconds: 400);
+const wordAnimationDuration = Duration(milliseconds: 1000);
